Refactor ProfileTabsList to simplify state management
Moved ProfileTabsList to a client-specific directory and removed unnecessary `useEffect` logic for managing tab state. Simplified tab handling by directly using `currentPath` for active tab tracking.
